chore: Add functionality for managing learners' attendance
This commit adds functionality to the application for managing learners' attendance in a training school. Users can now be added to the app and are assumed to be attending each class. If a user is unable to attend a class, they can go to the app and de-select themselves. Additionally, they have the option to join a different class if there is space available.